fix: useSearchParams()をSuspenseで囲んでビルドエラーを解消
Next.js 16では、useSearchParams()を使用する場合はSuspense境界で
囲む必要があるため、コンポーネントを分離して対応。
